pyproject.toml
[project]
name = "PGVectorScaleDB"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["python-dotenv>=1.0.1", "langchain-openai>=0.3.4", "timescale-vector>=0.0.7", "pandas>=2.2.3", "openai>=1.61.1", "instructor>=1.7.2", "anthropic>=0.45.2", "ipykernel>=6.29.5", "pip>=25.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm.build]
package-dir='src'

[tool.pdm]
distribution = false